refactor: Test Fixture 패키지 변경 및 가독성 향상
