+++ src/test/java/subway/TestFixture.java
@@ -1,4 +1,4 @@
-package subway.integration;
+package subway;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -15,12 +15,15 @@ public class TestFixture {
     public static final Line LINE_2 = new Line(1L, new LineName("2호선"));
     public static final Line LINE_3 = new Line(2L, new LineName("3호선"));
     public static final Line LINE_4 = new Line(3L, new LineName("4호선"));
+
     public static final Station JAMSIL = new Station(1L, "잠실");
     public static final Station GANGNAM = new Station(2L, "강남");
     public static final Station SEONGLENUG = new Station(3L, "선릉");
     public static final Station SAMSUNG = new Station(4L, "삼성");
+
     public static final Section SECTION_1 = new Section(1L, GANGNAM, SEONGLENUG, new Distance(5));
     public static final Section SECTION_2 = new Section(2L, SEONGLENUG, SAMSUNG, new Distance(5));
+
     public static final Sections SECTIONS = new EmptySections()
             .addSection(SECTION_1)
             .addSection(SECTION_2);
